feat: Setup HTML Webpack Plugin
HTML Webpack Plugin adds a reference to the /dist folder in index.html
so that a hardcoded reference to the JS file is needed. To setup HTML
Webpack Plugin, add a require statement at the beginning of the webpack
config file. Then tell wepback when to run the plugin. Update server.js
in server file accordingly (update folder in express.static to 'dist')